chore: extract common createIPFS function for tests
