+++ packages/core/src/__tests__/ceramic-api.test.ts
@@ -1,36 +1,16 @@
 import Ceramic, { CeramicConfig } from '../ceramic'
 import { Ed25519Provider } from 'key-did-provider-ed25519'
-import IPFS from 'ipfs-core'
 import tmp from 'tmp-promise'
 import { TileDoctype, TileParams } from "@ceramicnetwork/doctype-tile"
 import { AnchorStatus, DoctypeUtils, IpfsApi } from "@ceramicnetwork/common"
-
-import dagJose from 'dag-jose'
-import basicsImport from 'multiformats/cjs/src/basics-import.js'
-import legacy from 'multiformats/cjs/src/legacy.js'
 import DocID from '@ceramicnetwork/docid'
 import * as u8a from 'uint8arrays'
 import cloneDeep from 'lodash.clonedeep'
+import { createIPFS } from './create-ipfs';
 
 jest.mock('../store/level-state-store')
 
 const seed = u8a.fromString('6e34b2e1a9624113d81ece8a8a22e6e97f0e145c25c1d4d2d0e62753b4060c83', 'base16')
-
-/**
- * Create an IPFS instance
- * @param overrideConfig - IFPS config for override
- */
-const createIPFS =(overrideConfig: Record<string, unknown> = {}): Promise<IpfsApi> => {
-  basicsImport.multicodec.add(dagJose)
-  const format = legacy(basicsImport, dagJose.name)
-
-  const config = {
-    ipld: { formats: [format] },
-  }
-
-  Object.assign(config, overrideConfig)
-  return IPFS.create(config)
-}
 
 const registerChangeListener = function (doc: any): Promise<void> {
   return new Promise(resolve => {
@@ -49,6 +29,8 @@ const registerChangeListener = function (doc: any): Promise<void> {
  */
 const anchorDoc = async (ceramic: Ceramic, doc: any): Promise<void> => {
   const changeHandle = registerChangeListener(doc)
+  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+  // @ts-ignore
   await ceramic.context.anchorService.anchor()
   await changeHandle
 }
