+++ packages/core/src/__tests__/ceramic-pinning.test.ts
@@ -1,43 +1,12 @@
-import IPFS from 'ipfs-core'
 import Ceramic from '../ceramic'
 import { Ed25519Provider } from 'key-did-provider-ed25519'
 import tmp from 'tmp-promise'
 import getPort from 'get-port'
-import { DoctypeUtils, DocState, Doctype, IpfsApi } from "@ceramicnetwork/common"
-import { TileDoctype } from "@ceramicnetwork/doctype-tile"
-
-import dagJose from 'dag-jose'
-import basicsImport from 'multiformats/cjs/src/basics-import.js'
-import legacy from 'multiformats/cjs/src/legacy.js'
+import { IpfsApi, CeramicApi } from '@ceramicnetwork/common';
 import * as u8a from 'uint8arrays'
-
-//jest.mock('../store/level-state-store')
+import { createIPFS } from './create-ipfs';
 
 const seed = u8a.fromString('6e34b2e1a9624113d81ece8a8a22e6e97f0e145c25c1d4d2d0e62753b4060c83', 'base16')
-
-/**
- * Create an IPFS instance
- * @param overrideConfig - IPFS config for override
- */
-const createIPFS =(overrideConfig: Record<string, unknown> = {}): Promise<any> => {
-  basicsImport.multicodec.add(dagJose)
-  const format = legacy(basicsImport, dagJose.name)
-
-  const config = {
-    ipld: { formats: [format] },
-  }
-
-  Object.assign(config, overrideConfig)
-  return IPFS.create(config)
-}
-
-const expectEqualStates = (state1: DocState, state2: DocState): void => {
-  expect(DoctypeUtils.serializeState(state1)).toEqual(DoctypeUtils.serializeState(state2))
-}
-
-async function delay(mills: number): Promise<void> {
-  await new Promise(resolve => setTimeout(() => resolve(), mills))
-}
 
 const createCeramic = async (ipfs: IpfsApi, pinsetDirectory, anchorOnRequest = false): Promise<Ceramic> => {
   const ceramic = await Ceramic.create(ipfs, {
@@ -51,19 +20,7 @@ const createCeramic = async (ipfs: IpfsApi, pinsetDirectory, anchorOnRequest = f
   return ceramic
 }
 
-const anchor = async (ceramic: Ceramic): Promise<void> => {
-  await ceramic.context.anchorService.anchor()
-}
-
-const syncDoc = async (doctype: Doctype): Promise<void> => {
-  await new Promise(resolve => {
-    doctype.on('change', () => {
-      resolve()
-    })
-  })
-}
-
-const createDoc = async (ceramic: CeramicApi, controller: string, family: string): Promise<void> => {
+const createDoc = async (ceramic: CeramicApi, controller: string, family: string) => {
   return ceramic.createDocument(
     'tile',
     { deterministic: true, metadata: { controllers: [controller], family } },
@@ -77,20 +34,16 @@ describe('Ceramic document pinning', () => {
   let ipfs1: IpfsApi;
   let tmpFolder: any;
 
-  const DOCTYPE_TILE = 'tile'
-
   let p1Start = 4400
-
   const pOffset = 100
-
-  let port1: number;
 
   beforeEach(async () => {
     tmpFolder = await tmp.dir({ unsafeCleanup: true })
 
     const buildConfig = (path: string, port: number): Record<string, unknown> => {
       return {
-        repo: `${path}/ipfs${port}/`, config: {
+        repo: `${path}/ipfs${port}/`,
+        config: {
           Addresses: { Swarm: [`/ip4/127.0.0.1/tcp/${port}`] }, Bootstrap: []
         }
       }
@@ -100,7 +53,7 @@ describe('Ceramic document pinning', () => {
       return await getPort({port: getPort.makeRange(start + 1, start + offset)})
     }
 
-    ipfs1 = await createIPFS(buildConfig(tmpFolder.path), await findPort(p1Start, pOffset))
+    ipfs1 = await createIPFS(buildConfig(tmpFolder.path, await findPort(p1Start, pOffset)))
     p1Start += pOffset
   })
