+++ packages/core/src/__tests__/ceramic.test.ts
@@ -1,36 +1,16 @@
-import IPFS from 'ipfs-core'
 import Ceramic from '../ceramic'
 import { Ed25519Provider } from 'key-did-provider-ed25519'
 import tmp from 'tmp-promise'
 import getPort from 'get-port'
 import { DoctypeUtils, DocState, Doctype, IpfsApi } from "@ceramicnetwork/common"
 import { TileDoctype } from "@ceramicnetwork/doctype-tile"
-
-import dagJose from 'dag-jose'
-import basicsImport from 'multiformats/cjs/src/basics-import.js'
-import legacy from 'multiformats/cjs/src/legacy.js'
 import * as u8a from 'uint8arrays'
 import DocID from "@ceramicnetwork/docid"
+import { createIPFS } from './create-ipfs';
 
 jest.mock('../store/level-state-store')
 
 const seed = u8a.fromString('6e34b2e1a9624113d81ece8a8a22e6e97f0e145c25c1d4d2d0e62753b4060c83', 'base16')
-
-/**
- * Create an IPFS instance
- * @param overrideConfig - IPFS config for override
- */
-const createIPFS =(overrideConfig: Record<string, unknown> = {}): Promise<any> => {
-  basicsImport.multicodec.add(dagJose)
-  const format = legacy(basicsImport, dagJose.name)
-
-  const config = {
-    ipld: { formats: [format] },
-  }
-
-  Object.assign(config, overrideConfig)
-  return IPFS.create(config)
-}
 
 const expectEqualStates = (state1: DocState, state2: DocState): void => {
   expect(DoctypeUtils.serializeState(state1)).toEqual(DoctypeUtils.serializeState(state2))
